Update EventOntology.ttl
There are a few lines of differentiation we could choose to follow when representing these terms. Although the methods and vary, and they focus on two sets of animals, I think fishing is a form of hunting. Certainly, both can be done for recreation, commercial, or ecological regulation purposes. But are the methods and types of animals being targeted enough of a difference to make these sibling classes rather than parent-child classes?